requirements.txt
Flask==0.10.1
oauthlib==0.7.2
pycrypto==2.6.1
lxml==3.4.4
simplifycommerce-sdk-python==1.2.0
web.py==0.37
